修复视频发布接口内存泄露 bug；修改存储模块；修改数据库模块；新增限流中间件

+++ storage/oss.go
@@ -15,21 +15,53 @@ type OSS struct {
 func (*OSS) GetURL(path string) string {
 	return fmt.Sprintf(
 		"https://%s.%s/%s",
-		config.Conf.StorageConfig.OSS.BucketName,
-		config.Conf.StorageConfig.OSS.Endpoint,
+		config.Conf.Storage.OSS.BucketName,
+		config.Conf.Storage.OSS.Endpoint,
 		path,
 	)
 }
 
 func (o *OSS) Upload(path string, reader io.Reader) error {
-	return o.ossBucket.PutObject(
+	if err := o.ossBucket.PutObject(
 		path,
 		reader,
 		aliyunoss.ObjectACL(aliyunoss.ACLPublicRead),
-	)
+	); err != nil {
+		return UploadingError{err, path}
+	}
+
+	return nil
 }
 
-func (o *OSS) Delete(path ...string) ([]string, error) {
+func (o *OSS) Delete(path ...string) error {
+	if len(path) == 0 {
+		return nil
+	}
+
+	var deletingErr DeletingError
 	res, err := o.ossBucket.DeleteObjects(path)
-	return res.DeletedObjects, err
+	if err != nil {
+		deletingErr.errs = append(deletingErr.errs, err)
+		return deletingErr
+	}
+
+	numDeleted := len(res.DeletedObjects)
+	if numDeleted == len(path) {
+		return nil
+	}
+
+	// 找出删除失败的对象
+	deletedObjects := make(map[string]struct{}, numDeleted)
+	for _, key := range res.DeletedObjects {
+		deletedObjects[key] = struct{}{}
+	}
+	for _, p := range path {
+		if _, ok := deletedObjects[p]; !ok {
+			deletingErr.errs = append(
+				deletingErr.errs,
+				fmt.Errorf("删除对象 %s 失败", p),
+			)
+		}
+	}
+	return deletingErr
 }
